backend/requirements/local.txt
-r base.txt #includes the base.txt requirements file
arrow==0.10.0
Babel==2.5.1
binaryornot==0.4.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.4.1
Django==1.11.6
django-admin-honeypot==1.0.0
djangorestframework==3.7.1
docutils==0.14
future==0.16.0
idna==2.6
imagesize==0.7.1
Jinja2==2.9.6
jinja2-time==0.2.0
MarkupSafe==1.0
poyo==0.4.1
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
urllib3==1.22
virtualenv==15.1.0
whichcraft==0.4.1